+++ CHANGES.rst
@@ -21,6 +21,8 @@ Improvements
 - Use a more modern search dialog when searching for users (:issue:`4674`, :pr:`4743`)
 - Add an option to refresh event person data from the underlying user when cloning an
   event (:issue:`4750`, :pr:`4760`)
+- Add options for attaching iCal files to complete registration and event reminder
+  emails (:issue:`1158`, :pr:`4780`)
 
 Internal Changes
 ^^^^^^^^^^^^^^^^